Top Bar
The top bars for all of the pages except for the login page link to
pages that I thought were appropriate (Ones from this list: (Settings,
Create, Suggest, or Login) that could not be accessed directly from
that page ). The login page has a top bar that contains no links, only
a back arrow.