Improved ThemeActivity
Wrapped radio buttons into a radio group, so they don't need code to
switch between each other.  Made them call a function to set theme.
Variable theme stores the theme directly, so you don't have to look up
which it is, and so adding a new one is very simple.