Scrolling & Account blocks
I have tried to add a scroll to the textview & when you try to make an
account with blank textview / edittext's, it will deny the making of
your account.I can't tell if these functions are working because my
emulator never was able to work & my phone is buggy. So, can someone
please help test the app?